Create Project4: Treasure Island Project
My goal was to build a "Chose your own adventure game". Using If Else, Statements, Multiple Ifs, Nested and Elif statements I've learnt in python in this project. I built a very simple version of this type of text game and I was challenged to use python concepts on this game. 

PS: Tell me if you won! 

Signed-off-by: Thabile Jaku <107741511+jakuthab@users.noreply.github.com>